refactor(launch): split runtime_helpers into directory module
Splits crates/crosshook-core/src/launch/runtime_helpers.rs (1303 lines)
into a directory-backed module with per-responsibility submodules:

- proton_command — Proton/Gamescope command builders
- environment — env var merging and application
- path_resolution — Wine/Proton/Steam/working-dir paths
- platform — UMU discovery, Flatpak DBUS, launch capabilities

Plus a tests/ subdirectory mirroring the split.

No behavior change — pure structural split. Host-gateway contract
(ADR-0001) preserved: command construction still routes through
platform::host_command_with_env_and_directory*. Public API unchanged;
all existing callers across launch/, install/, onboarding/, prefix_deps/,
update/, export/, and request/ continue to resolve via
crate::launch::runtime_helpers::*.

Part of #290 (task 17).
