+++ src/crosshook-native/crates/crosshook-core/src/launch/runtime_helpers/tests/mod.rs
@@ -1,0 +1,4 @@
+mod path_resolution;
+mod platform;
+mod proton_command;
+mod support;
